code-example-tests/python/pymongo/requirements.txt
pylint==3.3.7
pymongo==4.13.2
python-dotenv==1.1.1
black
